fix: share token-refresh types and literal template helpers
