+++ packages/react-native-nitro-fetch/src/tokenRefresh.ts
@@ -1,4 +1,12 @@
 import { NativeStorage as NativeStorageSingleton } from './NitroInstances';
+import {
+  getNestedField,
+  applyTemplate,
+  applyCompositeTemplate,
+} from './tokenRefreshConfig';
+import type { TokenRefreshConfig } from './tokenRefreshConfig';
+export { getNestedField, applyTemplate } from './tokenRefreshConfig';
+export type { TokenRefreshConfig } from './tokenRefreshConfig';
 
 // Storage keys
 const KEY_WS = 'nitro_token_refresh_websocket';
@@ -7,79 +15,6 @@ const KEY_WS_CACHE = 'nitro_token_refresh_ws_cache';
 const KEY_FETCH_CACHE = 'nitro_token_refresh_fetch_cache';
 
 type TokenRefreshTarget = 'websocket' | 'fetch' | 'all';
-
-type TokenRefreshJsonMapping = {
-  jsonPath: string;
-  header: string;
-  valueTemplate?: string;
-};
-
-type TokenRefreshCompositeHeader = {
-  header: string;
-  template: string;
-  paths: Record<string, string>;
-};
-
-type TokenRefreshBodyMapping = {
-  jsonPath: string;
-  bodyPath: string;
-  valueTemplate?: string;
-};
-
-type TokenRefreshFormDataMapping = {
-  jsonPath: string;
-  field: string;
-  valueTemplate?: string;
-};
-
-export type TokenRefreshConfig = {
-  url: string;
-  method?: 'GET' | 'POST' | 'PUT' | 'PATCH';
-  headers?: Record<string, string>;
-  body?: string;
-  responseType?: 'json' | 'text';
-  mappings?: TokenRefreshJsonMapping[];
-  compositeHeaders?: TokenRefreshCompositeHeader[];
-  textHeader?: string;
-  textTemplate?: string;
-  bodyMappings?: TokenRefreshBodyMapping[];
-  formDataMappings?: TokenRefreshFormDataMapping[];
-  bodyTextPath?: string;
-  formDataTextField?: string;
-  onFailure?: 'skip' | 'useStoredHeaders';
-};
-
-// — Helpers —
-
-/**
- * Resolve a dot-notation path inside a parsed JSON object.
- */
-export function getNestedField(
-  obj: unknown,
-  dotPath: string
-): string | undefined {
-  const parts = dotPath.split('.');
-  let current: unknown = obj;
-  for (const part of parts) {
-    if (current == null || typeof current !== 'object') return undefined;
-    current = (current as Record<string, unknown>)[part];
-  }
-  return current != null ? String(current) : undefined;
-}
-
-export function applyTemplate(template: string, value: string): string {
-  return template.replace(/\{\{value\}\}/g, () => value);
-}
-
-function applyCompositeTemplate(
-  template: string,
-  values: Record<string, string>
-): string {
-  return template.replace(
-    /\{\{(\w+)\}\}/g,
-    (_: string, key: string) => values[key] ?? ''
-  );
-}
 
 export async function callRefreshEndpoint(
   config: TokenRefreshConfig
@@ -125,10 +60,9 @@ export async function callRefreshEndpoint(
 
   if (config.compositeHeaders) {
     for (const comp of config.compositeHeaders) {
-      const values: Record<string, string> = {};
+      const values: Record<string, string> = Object.create(null);
       for (const [placeholder, jsonPath] of Object.entries(comp.paths)) {
-        const val = getNestedField(json, jsonPath);
-        if (val != null) values[placeholder] = val;
+        values[placeholder] = getNestedField(json, jsonPath) ?? '';
       }
       headers[comp.header] = applyCompositeTemplate(comp.template, values);
     }
